refactoring imports bug fix
